src/requirements.txt
slackclient
packaging
